chore(protocol): move types
