+++ crates/protocol/comp/src/shadow.rs
