+++ crates/protocol/comp/src/traits.rs
