+++ crates/protocol/comp/src/variant.rs
