+++ crates/protocol/protocol/src/brotli.rs
